engine/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0
python-dotenv>=1.0.0
gunicorn==21.2.0
